move viewers to submodule
