+++ setup.py
@@ -43,6 +43,7 @@ setup(name='trimesh',
           'trimesh.path.exchange',
           'trimesh.scene',
           'trimesh.visual',
+          'trimesh.viewer',
           'trimesh.exchange',
           'trimesh.resources',
           'trimesh.interfaces'],
