+++ tests/test_html.py
@@ -8,18 +8,25 @@ class ViewTest(g.unittest.TestCase):
 
     def test_JSHTML(self):
 
-        import trimesh.scene.viewerJS as js
+        import trimesh.viewer.notebook as js
 
         m = g.get_mesh('featuretype.STL')
         s = m.scene()
 
         # a viewable scene
         html = js.scene_to_html(s)
-
         assert len(html) > 0
 
+        # check it out as an LXML document
+        from lxml.html import fromstring
+        h = fromstring(html)
+
+        # should have some children
+        children = list(h.body.iterchildren())
+        assert len(children) >= 2
+
     def test_inNB(self):
-        import trimesh.scene.viewerJS as js
+        import trimesh.viewer.notebook as js
 
         # should only return True inside
         # jypyter/IPython notebooks
