+++ trimesh/scene/__init__.py
@@ -2,7 +2,5 @@ from .camera import Camera
 
 from .scene import Scene, split_scene
 
-from .viewerJS import in_notebook
-
 # add to __all__ as per pep8
-__all__ = [Camera, Scene, split_scene, in_notebook]
+__all__ = [Camera, Scene, split_scene]
