+++ trimesh/viewer/notebook.py
