+++ trimesh/viewer/windowed.py
